feat: Implement all ZIL bit flags with 64-bit flag support
- Added all ZIL flags and search bits to ObjectFlag enum in flags.h
- Fixed flag casting in ZObject setFlag/clearFlag/hasFlag from uint32_t to uint64_t
- Updated getAllFlags/setAllFlags accessors to uint64_t
- Added MAZEBIT flag to all 15 maze rooms in world_init.cpp
- Added parser_flag_tests covering all 64-bit flags and bits
- Verified with parser_flag_tests and object_system_tests

+++ src/core/object.cpp
@@ -14,15 +14,15 @@ int ZObject::getProperty(PropertyId prop) const {
 }
 
 void ZObject::setFlag(ObjectFlag flag) {
-    flags_ |= static_cast<uint32_t>(flag);
+    flags_ |= static_cast<uint64_t>(flag);
 }
 
 void ZObject::clearFlag(ObjectFlag flag) {
-    flags_ &= ~static_cast<uint32_t>(flag);
+    flags_ &= ~static_cast<uint64_t>(flag);
 }
 
 bool ZObject::hasFlag(ObjectFlag flag) const {
-    return (flags_ & static_cast<uint32_t>(flag)) != 0;
+    return (flags_ & static_cast<uint64_t>(flag)) != 0;
 }
 
 void ZObject::moveTo(ZObject* location) {
